chore: Extend TrackableComponent default extension to dropin
